# dashboard/api/openapi.yaml
openapi: 3.1.0
...
info:
  title: MDDash Dashboard API
  version: 1.0.0
  description: API for experiment, simulation, compute, analysis, notebook, file, and publication workflows.
servers:
  - url: /
...
paths:
...
  /dash/api/experiments/{experiment_id}/analysis:
    parameters:
      - {$ref: '#/components/parameters/ExperimentId'}
    get:
      tags: [Analysis]
      summary: List analysis jobs
      operationId: listAnalysisJobs
      parameters:
        - {$ref: '#/components/parameters/SimulationPathQuery'}
      responses:
        '200':
          description: Analysis jobs
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/AnalysisJob'}}
        default: {$ref: '#/components/responses/Problem'}
    post:
      tags: [Analysis]
      summary: Submit an analysis job
      operationId: submitAnalysisJob
      requestBody:
        required: true
        content:
          application/json:
            schema: {$ref: '#/components/schemas/AnalysisJobRequest'}
      responses:
        '201':
          description: Submitted analysis job
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AnalysisJob'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/Problem'}
  /dash/api/experiments/{experiment_id}/analysis/types:
    get:
      tags: [Analysis]
      summary: List supported analysis types
      description: The hard set of mwf analysis task names derived from the MDDB workflow — the pool of analyses a simulation can produce.
      operationId: listAnalysisTypes
      parameters:
        - {$ref: '#/components/parameters/ExperimentId'}
      responses:
        '200':
          description: Analysis type ids
          content:
            application/json:
              schema:
                type: array
                items: {type: string}
        default: {$ref: '#/components/responses/Problem'}
  /dash/api/experiments/{experiment_id}/analysis/results:
    parameters:
      - {$ref: '#/components/parameters/ExperimentId'}
    get:
      tags: [Analysis]
      summary: List available analysis results
      operationId: listAnalysisResults
      parameters:
        - {$ref: '#/components/parameters/RequiredSimulationPathQuery'}
      responses:
        '200':
          description: Normalized analysis result names
          content:
            application/json:
              schema: {type: array, items: {type: string}}
        '400': {$ref: '#/components/responses/BadRequest'}
        default: {$ref: '#/components/responses/Problem'}
  /dash/api/experiments/{experiment_id}/analysis/results/{name}/variants:
    parameters:
      - {$ref: '#/components/parameters/ExperimentId'}
      - {$ref: '#/components/parameters/AnalysisResultName'}
    get:
      tags: [Analysis]
      summary: List variants for an analysis result
      operationId: listAnalysisResultVariants
      parameters:
        - {$ref: '#/components/parameters/RequiredSimulationPathQuery'}
      responses:
        '200':
          description: Analysis variants
          content:
            application/json:
              schema: {type: array, items: {$ref: '#/components/schemas/AnalysisVariant'}}
        '400': {$ref: '#/components/responses/BadRequest'}
        default: {$ref: '#/components/responses/Problem'}
  /dash/api/experiments/{experiment_id}/analysis/results/{name}:
    parameters:
      - {$ref: '#/components/parameters/ExperimentId'}
      - {$ref: '#/components/parameters/AnalysisResultName'}
    get:
      tags: [Analysis]
      summary: Get a typed analysis result
      operationId: getAnalysisResult
      parameters:
        - {$ref: '#/components/parameters/RequiredSimulationPathQuery'}
      responses:
        '200':
          description: Supported MWF analysis result
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AnalysisResult'}
        '400': {$ref: '#/components/responses/BadRequest'}
        '404': {$ref: '#/components/responses/NotFound'}
        '422': {$ref: '#/components/responses/UnprocessableEntity'}
        default: {$ref: '#/components/responses/Problem'}
  /dash/api/experiments/{experiment_id}/analysis/{job_id}:
    parameters:
      - {$ref: '#/components/parameters/ExperimentId'}
      - {$ref: '#/components/parameters/JobId'}
    get:
      tags: [Analysis]
      summary: Get an analysis job
      operationId: getAnalysisJob
      responses:
        '200':
          description: Analysis job
          content:
            application/json:
              schema: {$ref: '#/components/schemas/AnalysisJob'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/Problem'}
    delete:
      tags: [Analysis]
      summary: Delete an analysis job
      operationId: deleteAnalysisJob
      responses:
        '204': {$ref: '#/components/responses/NoContent'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/Problem'}
  /dash/api/experiments/{experiment_id}/analysis/{job_id}/logs:
    parameters:
      - {$ref: '#/components/parameters/ExperimentId'}
      - {$ref: '#/components/parameters/JobId'}
    get:
      tags: [Analysis]
      summary: Get analysis job logs
      operationId: getAnalysisJobLogs
      parameters:
        - {name: tail, in: query, schema: {type: integer, minimum: 1, maximum: 10000, default: 200}}
      responses:
        '200': {$ref: '#/components/responses/TextValue'}
        '404': {$ref: '#/components/responses/NotFound'}
        default: {$ref: '#/components/responses/Problem'}
components:
  parameters:
    ExperimentId: {name: experiment_id, in: path, required: true, schema: {type: string, minLength: 1}}
...
    JobId: {name: job_id, in: path, required: true, schema: {type: string, minLength: 1}}
    AnalysisResultName: {name: name, in: path, required: true, schema: {type: string, minLength: 1}}
    SimulationPathQuery: {name: simulation_path, in: query, schema: {type: string, minLength: 1}}
    RequiredSimulationPathQuery: {name: simulation_path, in: query, required: true, schema: {type: string, minLength: 1}}
  responses:
...
    TextValue:
      description: Text represented as a JSON string
      content:
        application/json:
          schema: {type: string}
    NoContent: {description: Operation completed successfully}
...
    Problem:
      description: RFC 9457 problem details
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
    BadRequest:
      description: Invalid request
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
    NotFound:
      description: Resource not found
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
    UnprocessableEntity:
      description: Resource could not be processed
      content:
        application/problem+json:
          schema: {$ref: '#/components/schemas/Problem'}
...
  schemas:
    Problem:
      type: object
      description: RFC 9457 problem details. The HTTP status is carried by the response status line.
      required: [type, title, detail]
      properties:
        type: {type: string, format: uri, pattern: '^urn:mddash:'}
        title: {type: string, minLength: 1}
        detail: {type: string, minLength: 1}
        solution: {type: string, minLength: 1}
      additionalProperties: false
...
    JobStatus: {type: string, enum: [UNKNOWN, PENDING, RUNNING, FINISHED, ERROR]}
...
    AnalysisJob:
      type: object
      required: [id, experiment_id, simulation_path, analysis_name, created_at, status]
      properties:
        id: {type: string}
        experiment_id: {type: string}
        simulation_path: {type: string}
        analysis_name:
          type: string
          enum: [apl, clusters, density, dist, energies, hbonds, inter, linter, lorder, pairwise, pca, perres, pockets, rgyr, rmsds, rmsf, sas, thickness, tmscore]
        created_at: {type: string, format: date-time}
        status: {$ref: '#/components/schemas/JobStatus'}
        sim_progress:
          type: number
          nullable: true
          description: Fraction (0-1) of the simulation's steps available when the analysis inputs were snapshotted; null when unknown.
      additionalProperties: true
    AnalysisJobRequest:
      type: object
      required: [simulation_path, analysis]
      properties:
        simulation_path: {type: string, minLength: 1}
        analysis: {type: string, enum: [apl, clusters, density, dist, energies, hbonds, inter, linter, lorder, pairwise, pca, perres, pockets, rgyr, rmsds, rmsf, sas, thickness, tmscore]}
        preprocessing_mode: {type: string, enum: [as_is, image, image_fit], default: as_is}
      additionalProperties: false
...
    AnalysisVariant:
      type: object
      required: [analysis, name]
      properties:
        analysis: {type: string, minLength: 1}
        name: {type: string, minLength: 1}
      additionalProperties: true
    NumberMatrix: {type: array, items: {type: array, items: {type: number}}}
...
    StatisticalData:
      type: object
      required: [average, stddev, min, max, data]
      properties:
        average: {type: number}
        stddev: {type: number}
        min: {type: number}
        max: {type: number}
        data: {type: array, items: {type: number}}
      additionalProperties: false
    StatisticalSeriesResult:
      type: object
      required: [step, y]
      properties:
        start: {type: number}
        step: {type: number}
        y:
          type: object
          minProperties: 1
          additionalProperties: {$ref: '#/components/schemas/StatisticalData'}
      additionalProperties: false
    NamedSeriesResult:
      type: object
      required: [start, step, data]
      properties:
        start: {type: number}
        step: {type: number}
        data:
          type: array
          items:
            type: object
            required: [reference, group, values]
            properties:
              reference: {type: string}
              group: {type: string}
              values: {type: array, items: {type: number}}
            additionalProperties: false
      additionalProperties: false
    NamedDataResult:
      type: object
      required: [data]
      properties:
        step: {type: number}
        data:
          type: array
          items:
            type: object
            required: [name]
            properties: {name: {type: string}}
            additionalProperties: true
      additionalProperties: false
    MatrixResult:
      type: object
      properties:
        start: {type: number}
        step: {type: number}
        rmsdpr: {$ref: '#/components/schemas/NumberMatrix'}
        saspf:
          type: array
          items: {type: array, items: {type: [number, 'null']}}
        means: {type: array, items: {type: number}}
        stdvs: {type: array, items: {type: number}}
      anyOf: [{required: [rmsdpr]}, {required: [saspf]}]
      additionalProperties: false
    PCAResult:
      type: object
      required: [framestep, atoms, eigenvalues, projections]
      properties:
        framestep: {type: number}
        atoms: {type: array, items: {type: integer}}
        eigenvalues: {type: array, items: {type: number}}
        projections: {$ref: '#/components/schemas/NumberMatrix'}
      additionalProperties: false
    ClustersResult:
      type: object
      required: [name, cutoff, clusters, transitions, step, version]
      properties:
        name: {type: string}
        cutoff: {type: number}
        clusters:
          type: array
          items:
            type: object
            required: [frames, main]
            properties:
              frames: {type: array, items: {type: integer}}
              main: {type: integer}
            additionalProperties: false
        transitions:
          type: array
          items:
            type: object
            required: [from, to, count]
            properties:
              from: {type: integer}
              to: {type: integer}
              count: {type: integer}
            additionalProperties: false
        step: {type: number}
        version: {type: string}
      additionalProperties: false
    MembraneMapResult:
      type: object
      required: [n_mems, mems, no_mem_lipid]
      properties:
        n_mems: {type: integer}
        mems: {type: object, additionalProperties: {type: object}}
        no_mem_lipid: {type: array, items: {type: integer}}
      additionalProperties: false
    ThicknessResult:
      type: object
      required: [data]
      properties:
        step: {type: number}
        data:
          type: object
          required: [frame, mean_positive, mean_negative, std_positive, std_negative, thickness, std_thickness, midplane_z]
          additionalProperties: {type: array, items: {type: number}}
      additionalProperties: false
    InteractionResult:
      type: object
      required: [name, agent_1, agent_2]
      properties:
        name: {type: string}
        agent_1: {type: string}
        agent_2: {type: string}
        has_cg: {type: boolean}
        version: {type: string}
      additionalProperties: true
    AnalysisResult:
      description: Closed catalog of analysis result shapes currently supported by the dashboard.
      oneOf:
        - {$ref: '#/components/schemas/StatisticalSeriesResult'}
        - {$ref: '#/components/schemas/NamedSeriesResult'}
        - {$ref: '#/components/schemas/NamedDataResult'}
        - {$ref: '#/components/schemas/MatrixResult'}
        - {$ref: '#/components/schemas/PCAResult'}
        - {$ref: '#/components/schemas/ClustersResult'}
        - {$ref: '#/components/schemas/MembraneMapResult'}
        - {$ref: '#/components/schemas/ThicknessResult'}
        - {$ref: '#/components/schemas/InteractionResult'}